+++ packages/create/owc-app/src/owc-app.js
@@ -1,0 +1,3 @@
+import { OwcApp } from './OwcApp.js';
+
+customElements.define('owc-app', OwcApp);