+++ packages/create/src/generators/demoing-storybook/templates/static/.storybook/preview.js
@@ -1,20 +1,6 @@
-import {
-  addParameters,
-  addDecorator,
-  setCustomElements,
-  withA11y,
-} from '@open-wc/demoing-storybook';
-
-addDecorator(withA11y);
+import { addParameters, setCustomElements } from '@open-wc/demoing-storybook';
 
 addParameters({
-  a11y: {
-    config: {},
-    options: {
-      checks: { 'color-contrast': { options: { noScroll: true } } },
-      restoreScroll: true,
-    }
-  },
   docs: {
     iframeHeight: '200px',
   }