+++ packages/rollup-plugin-index-html/test/snapshots/js-entry/app.js
@@ -1,4 +1,0 @@
-// @ts-nocheck
-/* eslint-disable */
-console.log('my app');
-import('./lazy-ad0cfec5.js');