+++ packages/rollup-plugin-index-html/test/snapshots/legacy/app.js
@@ -1,5 +1,0 @@
-console.log('shared');
-
-console.log('my app');
-
-import('./lazy-2ca8ef13.js');