+++ packages/testing/test/chaiA11yAxe.test.js
@@ -3,16 +3,12 @@ import { fixture, expect, assert, html } from '../index.js';
 describe('chaiA11yAxe', () => {
   describe('Expect', () => {
     it('passes axe accessible tests', async () => {
-      const el = await fixture(html`
-        <button>some light dom</button>
-      `);
+      const el = await fixture(html` <button>some light dom</button> `);
       await expect(el).to.be.accessible();
     });
 
     it('accepts "done" option', done => {
-      fixture(html`
-        <button>some light dom</button>
-      `).then(el => {
+      fixture(html` <button>some light dom</button> `).then(el => {
         expect(el).to.be.accessible({
           done,
         });
@@ -31,34 +27,26 @@ describe('chaiA11yAxe', () => {
     });
 
     it('uses negation to pass failed test', async () => {
-      const el = await fixture(html`
-        <div aria-labelledby="test-x"></div>
-      `);
+      const el = await fixture(html` <div aria-labelledby="test-x"></div> `);
       await expect(el).not.to.be.accessible();
     });
   });
 
   describe('Assert', () => {
     it('passes axe accessible tests', async () => {
-      const el = await fixture(html`
-        <button>some light dom</button>
-      `);
+      const el = await fixture(html` <button>some light dom</button> `);
       await assert.isAccessible(el);
     });
 
     it('accepts ignored rules list', async () => {
-      const el = await fixture(html`
-        <div aria-labelledby="test-x"></div>
-      `);
+      const el = await fixture(html` <div aria-labelledby="test-x"></div> `);
       await assert.isAccessible(el, {
         ignoredRules: ['aria-valid-attr-value'],
       });
     });
 
     it('throws when audit did not pass', async () => {
-      const el = await fixture(html`
-        <div aria-labelledby="test-x"></div>
-      `);
+      const el = await fixture(html` <div aria-labelledby="test-x"></div> `);
       let thrown = false;
       try {
         await assert.isAccessible(el);
@@ -69,9 +57,7 @@ describe('chaiA11yAxe', () => {
     });
 
     it('passes for negation', async () => {
-      const el = await fixture(html`
-        <div aria-labelledby="test-x"></div>
-      `);
+      const el = await fixture(html` <div aria-labelledby="test-x"></div> `);
       await assert.isNotAccessible(el);
     });
   });