+++ packages/building-utils/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.16.4](https://github.com/open-wc/open-wc/compare/@open-wc/building-utils@2.16.3...@open-wc/building-utils@2.16.4) (2020-04-20)
+
+**Note:** Version bump only for package @open-wc/building-utils
+
+
+
+
+
 ## [2.16.3](https://github.com/open-wc/open-wc/compare/@open-wc/building-utils@2.16.2...@open-wc/building-utils@2.16.3) (2020-04-12)
 
 **Note:** Version bump only for package @open-wc/building-utils