+++ packages/polyfills-loader/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.6](https://github.com/open-wc/open-wc/compare/polyfills-loader@1.5.5...polyfills-loader@1.5.6) (2020-04-21)
+
+
+### Bug Fixes
+
+* **polyfills-loader:** allow forcing the polyfill ([ee44812](https://github.com/open-wc/open-wc/commit/ee44812f89e90879713a0ba5cd79e631c57f6d4e))
+
+
+
+
+
+## [1.5.5](https://github.com/open-wc/open-wc/compare/polyfills-loader@1.5.4...polyfills-loader@1.5.5) (2020-04-20)
+
+**Note:** Version bump only for package polyfills-loader
+
+
+
+
+
 ## [1.5.4](https://github.com/open-wc/open-wc/compare/polyfills-loader@1.5.3...polyfills-loader@1.5.4) (2020-04-12)
 
 **Note:** Version bump only for package polyfills-loader